requirements.txt
python-telegram-bot==20.3
Flask==2.3.3
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0